Make money commands atomic and idempotent with durable outboxes and reporting projections

+++ microservices/account-service/src/main/java/com/bellagnech/account/dtos/BankAccountDTO.java
@@ -1,4 +1,5 @@
 package com.bellagnech.account.dtos;
+import java.math.BigDecimal;
 
 import com.bellagnech.account.enums.AccountStatus;
 import lombok.AllArgsConstructor;
@@ -12,14 +13,14 @@ import java.util.Date;
 @AllArgsConstructor
 public class BankAccountDTO {
     private String id;
-    private double balance;
+    private BigDecimal balance;
     private Date createDate;
     private AccountStatus status;
     private Long customerId;
     private String type;
     private String customerName;
     private String customerEmail;
-    private Double overDraft;   // for current accounts
+    private BigDecimal overDraft;   // for current accounts
     private Double interestRate; // for saving accounts
 }
