+++ microservices/transaction-service/src/main/java/com/bellagnech/transaction/dtos/TransactionRequest.java
@@ -1,4 +1,5 @@
 package com.bellagnech.transaction.dtos;
+import java.math.BigDecimal;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -13,11 +14,11 @@ import lombok.NoArgsConstructor;
 public class TransactionRequest {
     @NotBlank(message = "Account ID is required")
     private String accountId;
-    
+
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private Double amount;
-    
+    private BigDecimal amount;
+
     @NotBlank(message = "Description is required")
     private String description;
 }
