+++ microservices/transaction-service/src/main/java/com/bellagnech/transaction/dtos/TransferRequest.java
@@ -1,4 +1,5 @@
 package com.bellagnech.transaction.dtos;
+import java.math.BigDecimal;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -13,14 +14,14 @@ import lombok.NoArgsConstructor;
 public class TransferRequest {
     @NotBlank(message = "Source account ID is required")
     private String sourceAccountId;
-    
+
     @NotBlank(message = "Destination account ID is required")
     private String destinationAccountId;
-    
+
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private Double amount;
-    
+    private BigDecimal amount;
+
     private String description;
 }
